+++ project/src/pages/login-screen/login-screen.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/header/header';
+import Header from '../../components/header/header';
 
 function LoginScreen() {
   return (
